Moved PMTconfiguration object from icarus to sbn namespace.
Source code moved accordingly. And some comments fixed.
