+++ sbnobj/Common/CMakeLists.txt
@@ -1,4 +1,5 @@
 add_subdirectory(CRT)
+add_subdirectory(PMT)
 add_subdirectory(Reco)
 add_subdirectory(Analysis)
