+++ sbnobj/Common/PMT/Data/CMakeLists.txt
@@ -1,0 +1,9 @@
+cet_make(
+  LIBRARIES
+  NO_DICTIONARY
+  )
+
+art_dictionary(DICTIONARY_LIBRARIES sbnobj_Common_PMT_Data)
+
+install_headers()
+install_source()
