+++ sbnobj/Common/PMT/Data/V1730channelConfiguration.h
@@ -1,39 +1,38 @@
 /**
- * @file   sbnobj/ICARUS/PMT/Data/V1730Configuration.h
+ * @file   sbnobj/Common/PMT/Data/V1730channelConfiguration.h
  * @brief  Information from the configuration of a V1730 PMT readout board.
  * @author Gianluca Petrillo (petrillo@slac.stanford.edu)
  * @date   February 18, 2021
- * @see    sbnobj/ICARUS/PMT/Data/V1730Configuration.cxx
+ * @see    sbnobj/Common/PMT/Data/V1730channelConfiguration.cxx
  */
 
-#ifndef SBNOBJ_ICARUS_PMT_DATA_V1730CONFIGURATION_H
-#define SBNOBJ_ICARUS_PMT_DATA_V1730CONFIGURATION_H
+#ifndef SBNOBJ_COMMON_PMT_DATA_V1730CHANNELCONFIGURATION_H
+#define SBNOBJ_COMMON_PMT_DATA_V1730CHANNELCONFIGURATION_H
 
-
-// ICARUS libraries
-#include "sbnobj/ICARUS/PMT/Data/V1730channelConfiguration.h"
+// LArSoft libraries
+#include "lardataobj/RawData/OpDetWaveform.h" // raw::Channel_t
 
 // C/C++ standard libraries
 #include <iosfwd> // std::ostream
-#include <vector>
 #include <string>
+#include <limits>
 
 
 //------------------------------------------------------------------------------
-namespace icarus {
+namespace sbn {
   
-  struct V1730Configuration;
+  struct V1730channelConfiguration;
   
   /// Prints the configuration into a stream with default verbosity.
   std::ostream& operator<<
-    (std::ostream& out, icarus::V1730Configuration const& config);
+    (std::ostream& out, sbn::V1730channelConfiguration const& config);
   
-} // namespace icarus
+} // namespace sbn
 
 /**
- * @brief Class containing configuration for a V1730 board.
+ * @brief Class containing configuration for a V1730 channel.
  * 
- * This is an informative class containing configuration of a V1730 board
+ * This is an informative class containing configuration of a V1730 channel
  * extracted from some other source (typically, DAQ) made readily available
  * to the users.
  * 
@@ -43,52 +42,41 @@ namespace icarus {
  * element by element.
  * 
  */
-struct icarus::V1730Configuration {
+struct sbn::V1730channelConfiguration {
+  
+  /// Special value for unassigned channel ID.
+  static constexpr auto NoChannelID
+    = std::numeric_limits<raw::Channel_t>::max();
   
   // --- BEGIN -- Data members -------------------------------------------------
   
   // NOTE when adding data members, remember to add an element to the comparison
   
-  /// Name (mnemonic) of the board.
-  std::string boardName;
+  /// Number of the channel on the board (0-15).
+  short unsigned int channelNo = std::numeric_limits<short unsigned int>::max();
   
-  /// Numeric ID of the board (`board_id`).
-  unsigned int boardID = std::numeric_limits<unsigned int>::max();
+  /// Offline channel ID.
+  raw::Channel_t channelID = NoChannelID;
   
-  /// DAQ fragment ID.
-  unsigned int fragmentID = std::numeric_limits<unsigned int>::max();
+  /// Baseline (`BaselineCh<N+1>`).
+  short signed int baseline = 0;
   
-  /// Ticks in each buffer (`recordLength`).
-  unsigned int bufferLength = 0U;
+  /// Threshold (`triggerThreshold<N>`).
+  short signed int threshold = 0;
   
-  /// Fraction of the waveform _after_ the trigger signal (`postPercent`).
-  float postTriggerFrac = 0.0f;
-  
-  /// Number of channels (`nChannels`).
-  unsigned int nChannels = 0U;
-  
-  /// Configuration of each channel.
-  std::vector<icarus::V1730channelConfiguration> channels;
+  /// Channel is enabled (`enable`).
+  bool enabled = false;
   
   // --- END ---- Data members -------------------------------------------------
   
   
   // --- BEGIN -- Derived quantities -------------------------------------------
   
-  /// Ticks in the waveform before the trigger.
-  unsigned int preTriggerTicks() const;
+  /// Returns whether the channel ID is set.
+  bool hasChannelID() const;
   
-  /// Ticks in the waveform after the trigger.
-  unsigned int postTriggerTicks() const;
-  
-  /// Duration of the waveform [us].
-  float bufferTime() const;
-  
-  /// Time in the waveform before the trigger [us].
-  float preTriggerTime() const;
-  
-  /// Time in the waveform after the trigger [us].
-  float postTriggerTime() const;
+  /// Threshold relative to the baseline (ticks).
+  short signed int relativeThreshold() const;
   
   // --- END ---- Derived quantities -------------------------------------------
   
@@ -96,21 +84,20 @@ struct icarus::V1730Configuration {
 #if __cplusplus < 202004L
   //@{
   /// Comparison: all fields need to have the same values.
-  bool operator== (V1730Configuration const& other) const;
-  bool operator!= (V1730Configuration const& other) const
+  bool operator== (V1730channelConfiguration const& other) const;
+  bool operator!= (V1730channelConfiguration const& other) const
     { return ! this->operator== (other); }
   //@}
 #else
 # error "With C++20 support, enable the default comparison operators"
   // probably the compiler will be generating these anyway, so don't bother
-//   bool operator== (V1730Configuration const& other) const = default;
-//   bool operator!= (V1730Configuration const& other) const = default;
+//   bool operator== (V1730channelConfiguration const& other) const = default;
+//   bool operator!= (V1730channelConfiguration const& other) const = default;
 #endif
   
   // -- BEGIN -- Dump facility -------------------------------------------------
   /// Maximum supported verbosity level supported by `dump()`.
-  static constexpr unsigned int MaxDumpVerbosity
-    = V1730channelConfiguration::MaxDumpVerbosity + 1U;
+  static constexpr unsigned int MaxDumpVerbosity = 1U;
   
   /// Default verbosity level for `dump()`.
   static constexpr unsigned int DefaultDumpVerbosity = MaxDumpVerbosity;
@@ -129,11 +116,9 @@ struct icarus::V1730Configuration {
    * 
    * The amount of information printed depends on the `verbosity` level:
    * 
-   * * `0`: name of the board, board and fragment ID, number of channels
-   * * `1`: waveform length and post-trigger fraction;
-   * *   information on each channel, one per line, as with
-   * *   `V1730channelConfiguration::dump() with verbosity one level smaller
-   * *   than the value of `verbosity` argument
+   * * `0`: channel number and whether it is enabled or not (also off-line
+   *     channel ID if available)
+   * * `1`: also baseline and threshold
    * 
    */
   void dump(std::ostream& out,
@@ -170,64 +155,43 @@ struct icarus::V1730Configuration {
   
   // -- END ---- Dump facility -------------------------------------------------
   
-  
-  /// Convert a number of ticks into a time interval in microseconds.
-  static float ticks2us(int ticks) { return ticks * 0.002f; }
-
-}; // icarus::V1730Configuration
+}; // sbn::V1730channelConfiguration
 
 
 
 //------------------------------------------------------------------------------
 //---  Inline implementation
 //------------------------------------------------------------------------------
-inline unsigned int icarus::V1730Configuration::preTriggerTicks() const
-  { return static_cast<unsigned int>((1.0f - postTriggerFrac) * bufferLength); }
+inline bool sbn::V1730channelConfiguration::hasChannelID() const
+  { return channelID != NoChannelID; }
 
 
 //------------------------------------------------------------------------------
-inline unsigned int icarus::V1730Configuration::postTriggerTicks() const
-  { return static_cast<unsigned int>(postTriggerFrac * bufferLength); }
+inline short signed int sbn::V1730channelConfiguration::relativeThreshold
+  () const
+  { return baseline - threshold; }
 
 
 //------------------------------------------------------------------------------
-inline float icarus::V1730Configuration::bufferTime() const
-  { return ticks2us(bufferLength); }
-
-
-//------------------------------------------------------------------------------
-inline float icarus::V1730Configuration::preTriggerTime() const
-  { return ticks2us(preTriggerTicks()); }
-
-
-//------------------------------------------------------------------------------
-inline float icarus::V1730Configuration::postTriggerTime() const
-  { return ticks2us(postTriggerTicks()); }
-
-
-//------------------------------------------------------------------------------
-inline bool icarus::V1730Configuration::operator==
-  (icarus::V1730Configuration const& other) const
+inline bool sbn::V1730channelConfiguration::operator==
+  (sbn::V1730channelConfiguration const& other) const
 {
-  
-  if (boardName       != other.boardName      ) return false;
-  if (boardID         != other.boardID        ) return false;
-  if (fragmentID      != other.fragmentID     ) return false;
-  if (bufferLength    != other.bufferLength   ) return false;
-  if (postTriggerFrac != other.postTriggerFrac) return false;
-  if (nChannels       != other.nChannels      ) return false;
-  if (channels        != other.channels       ) return false;
+  if (channelNo != other.channelNo) return false;
+  if (channelID != other.channelID) return false;
+  if (baseline  != other.baseline ) return false;
+  if (threshold != other.threshold) return false;
+  if (enabled   != other.enabled  ) return false;
   
   return true;
-} // icarus::V1730Configuration::operator==()
+} // sbn::V1730channelConfiguration::operator==()
 
 
 //------------------------------------------------------------------------------
-inline std::ostream& icarus::operator<<
-  (std::ostream& out, icarus::V1730Configuration const& config)
+inline std::ostream& sbn::operator<<
+  (std::ostream& out, sbn::V1730channelConfiguration const& config)
   { config.dump(out); return out; }
 
 
 //------------------------------------------------------------------------------
 
-#endif // SBNOBJ_ICARUS_PMT_DATA_V1730CONFIGURATION_H
+#endif // SBNOBJ_COMMON_PMT_DATA_V1730CHANNELCONFIGURATION_H
