+++ sbnobj/ICARUS/PMT/Data/classes.h
@@ -1,5 +1,5 @@
 /**
- * @file   icaruscode/PMT/Data/classes.h
+ * @file   sbnobj/ICARUS/PMT/Data/classes.h
  * @author Gianluca Petrillo (petrillo@slac.stanford.edu)
  * @date   September 11, 2020
  * 
@@ -8,12 +8,11 @@
  * * `icarus::WaveformBaseline`
  *   (and its associations with `raw::OpDetWaveform`)
  * 
- * See also `icaruscode/PMT/Data/classes_def.xml`.
+ * See also `sbnobj/ICARUS/PMT/Data/classes_def.xml`.
  */
 
 // ICARUS libraries
 #include "sbnobj/ICARUS/PMT/Data/WaveformBaseline.h"
-#include "sbnobj/ICARUS/PMT/Data/PMTconfiguration.h"
 
 // LArSoft libraries
 #include "lardataobj/RawData/OpDetWaveform.h"
@@ -28,6 +27,5 @@
 namespace {
   
   icarus::WaveformBaseline baseline;
-  icarus::PMTconfiguration PMTconfig;
   
 } // local namespace
